# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PesaMatrix Copy Trading SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/dashboard:
    get:
      operationId: adminDashboard
      tags: [admin]
      summary: Admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user detail
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/audit-logs:
    get:
      operationId: adminAuditLogs
      tags: [admin]
      summary: Audit log trail
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogsResponse"
...
  /admin/subscriptions:
    get:
      operationId: adminListSubscriptions
      tags: [admin]
      summary: List all subscriptions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSubscriptionsResponse"
  /admin/plans:
    get:
      operationId: adminListPlans
      tags: [admin]
      summary: List all plans
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlansResponse"
    post:
      operationId: adminCreatePlan
      tags: [admin]
      summary: Create a subscription plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
  /admin/plans/{id}:
    patch:
      operationId: adminUpdatePlan
      tags: [admin]
      summary: Update a subscription plan
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get admin platform settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettingsResponse"
    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update admin platform settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettingsResponse"

  # ── Master account (admin) ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [admin, subscriber]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, email, role]
...
    Plan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        pricePerDay:
          type: string
        minimumDays:
          type: integer
        maximumDays:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, pricePerDay, minimumDays, maximumDays, features, isActive]
    PlansResponse:
      type: object
      properties:
        plans:
          type: array
          items:
            $ref: "#/components/schemas/Plan"
      required: [plans]
    PlanResponse:
      type: object
      properties:
        plan:
          $ref: "#/components/schemas/Plan"
      required: [plan]
    PlanInput:
      type: object
      properties:
        name:
          type: string
        pricePerDay:
          type: string
        minimumDays:
          type: integer
        maximumDays:
          type: integer
        features:
          type: array
          items:
            type: string
      required: [name, pricePerDay, minimumDays, maximumDays]
    PlanUpdate:
      type: object
      properties:
        name:
          type: string
        pricePerDay:
          type: string
        minimumDays:
          type: integer
        maximumDays:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean

    # ── Subscriptions ────────────────────────────────────────────────────────────
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, active, expired, cancelled]
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        numberOfDays:
          type: integer
        amountPaid:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, planId, status, isActive, numberOfDays, amountPaid]
...
    AdminSubscriptionsResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        total:
          type: integer
        active:
          type: integer
        expired:
          type: integer
      required: [subscriptions, total, active, expired]

    # ── Payments ────────────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        adminUsers:
          type: integer
        subscriberUsers:
          type: integer
        activeSubscriptions:
          type: integer
        totalRevenue:
          type: number
        pendingPayments:
          type: integer
      required: [totalUsers, adminUsers, subscriberUsers, activeSubscriptions, totalRevenue, pendingPayments]
    DashboardResponse:
      type: object
      properties:
        stats:
          $ref: "#/components/schemas/DashboardStats"
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        recentActivity:
          type: array
          items:
            type: object
      required: [stats, recentUsers, recentActivity]
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
      required: [users, total]
    UserDetailResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        subscription:
          type: ["object", "null"]
          $ref: "#/components/schemas/Subscription"
        activity:
          type: array
          items:
            type: object
      required: [user, activity]
    AuditLogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            type: object
      required: [logs]
...
    AdminSettingsResponse:
      type: object
      properties:
        subscriptionFeePerDay:
          type: string
        minimumSubscriptionDays:
          type: integer
        maximumSubscriptionDays:
          type: integer
        copyFactoryStrategyId:
          type: ["string", "null"]
        masterMt5Login:
          type: ["string", "null"]
      required: [subscriptionFeePerDay, minimumSubscriptionDays, maximumSubscriptionDays]
    AdminSettingsUpdate:
      type: object
      properties:
        subscriptionFeePerDay:
          type: string
        minimumSubscriptionDays:
          type: integer
        maximumSubscriptionDays:
          type: integer
...