# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PesaMatrix Copy Trading SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettingsUpdate {
  subscriptionFeePerDay?: string;
  minimumSubscriptionDays?: number;
  maximumSubscriptionDays?: number;
}