# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PesaMatrix Copy Trading SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  name: string;
  pricePerDay: string;
  minimumDays: number;
  maximumDays: number;
  features: string[];
  isActive: boolean;
  createdAt?: Date;
}