# lib/api-zod/src/generated/types/usersListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PesaMatrix Copy Trading SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface UsersListResponse {
  users: UserProfile[];
  total: number;
}